+++ crates/core_simd/src/vector/uint.rs
@@ -1,12 +1,40 @@
 #![allow(non_camel_case_types)]
 
+
+/// Implements additional integer traits (Eq, Ord, Hash) on the specified vector `$name`, holding multiple `$lanes` of `$type`.
+macro_rules! impl_unsigned_vector {
+    { $name:ident, $type:ty } => {
+        impl_vector! { $name, $type }
+
+        impl<const LANES: usize> Eq for $name<LANES> where Self: crate::LanesAtMost64 {}
+
+        impl<const LANES: usize> Ord for $name<LANES> where Self: crate::LanesAtMost64 {
+            #[inline]
+            fn cmp(&self, other: &Self) -> core::cmp::Ordering {
+                // TODO use SIMD cmp
+                self.to_array().cmp(other.as_ref())
+            }
+        }
+
+        impl<const LANES: usize> core::hash::Hash for $name<LANES> where Self: crate::LanesAtMost64 {
+            #[inline]
+            fn hash<H>(&self, state: &mut H)
+            where
+                H: core::hash::Hasher
+            {
+                self.as_slice().hash(state)
+            }
+        }
+    }
+}
+
 /// A SIMD vector of containing `LANES` `usize` values.
 #[repr(simd)]
 pub struct SimdUsize<const LANES: usize>([usize; LANES])
 where
     Self: crate::LanesAtMost64;
 
-impl_integer_vector! { SimdUsize, usize }
+impl_unsigned_vector! { SimdUsize, usize }
 
 #[cfg(target_pointer_width = "32")]
 from_transmute_x86! { unsafe usizex4 => __m128i }
@@ -26,7 +54,7 @@ pub struct SimdU128<const LANES: usize>([u128; LANES])
 where
     Self: crate::LanesAtMost64;
 
-impl_integer_vector! { SimdU128, u128 }
+impl_unsigned_vector! { SimdU128, u128 }
 
 from_transmute_x86! { unsafe u128x2 => __m256i }
 //from_transmute_x86! { unsafe u128x4 => __m512i }
@@ -37,7 +65,7 @@ pub struct SimdU16<const LANES: usize>([u16; LANES])
 where
     Self: crate::LanesAtMost64;
 
-impl_integer_vector! { SimdU16, u16 }
+impl_unsigned_vector! { SimdU16, u16 }
 
 from_transmute_x86! { unsafe u16x8 => __m128i }
 from_transmute_x86! { unsafe u16x16 => __m256i }
@@ -49,7 +77,7 @@ pub struct SimdU32<const LANES: usize>([u32; LANES])
 where
     Self: crate::LanesAtMost64;
 
-impl_integer_vector! { SimdU32, u32 }
+impl_unsigned_vector! { SimdU32, u32 }
 
 from_transmute_x86! { unsafe u32x4 => __m128i }
 from_transmute_x86! { unsafe u32x8 => __m256i }
@@ -61,7 +89,7 @@ pub struct SimdU64<const LANES: usize>([u64; LANES])
 where
     Self: crate::LanesAtMost64;
 
-impl_integer_vector! { SimdU64, u64 }
+impl_unsigned_vector! { SimdU64, u64 }
 
 from_transmute_x86! { unsafe u64x2 => __m128i }
 from_transmute_x86! { unsafe u64x4 => __m256i }
@@ -73,7 +101,7 @@ pub struct SimdU8<const LANES: usize>([u8; LANES])
 where
     Self: crate::LanesAtMost64;
 
-impl_integer_vector! { SimdU8, u8 }
+impl_unsigned_vector! { SimdU8, u8 }
 
 from_transmute_x86! { unsafe u8x16 => __m128i }
 from_transmute_x86! { unsafe u8x32 => __m256i }
